requirements.txt
cffi==1.11.2
click==6.7
cryptography==1.7.2
enum34==1.1.6
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.3.2
pyasn1==0.3.7
pycparser==2.18
PyJWT==1.5.3
six==1.11.0
SQLAlchemy==1.1.15
Werkzeug==0.12.2
